+++ Plugin.cs
@@ -1,12 +1,14 @@
 ﻿using System.Collections.Generic;
 using BepInEx;
 using System.IO;
+using System.Linq;
 using System.Reflection;
 using BepInEx.Configuration;
 using HarmonyLib;
 using LethalConfig;
 using LethalConfig.ConfigItems;
 using LethalConfig.ConfigItems.Options;
+using LethalLib.Modules;
 using SurprisingHazards.Patches;
 using SurprisingHazards.Scripts;
 using Unity.Netcode;
@@ -22,7 +24,7 @@ using UnityEngine;
 
         const string GUID = "wexop.surprising_hazards";
         const string NAME = "SurprisingHazards";
-        const string VERSION = "1.0.0";
+        const string VERSION = "1.0.1";
 
         public static SurprisingHazardsPlugin instance;
 
@@ -30,6 +32,7 @@ using UnityEngine;
         public Dictionary<ulong, RegisteredHazard> RegisteredHazards = new Dictionary<ulong, RegisteredHazard>();
 
         public ConfigEntry<float> visibleRange;
+        public ConfigEntry<string> customVisibleRange;
 
         void Awake()
         {
@@ -47,10 +50,19 @@ using UnityEngine;
             
             Debug.Log(surprisingGameObject.name);
             
+            
+            
             Harmony.CreateAndPatchAll(typeof(TurretPatch));
             Harmony.CreateAndPatchAll(typeof(LandminePatch));
             Harmony.CreateAndPatchAll(typeof(SpikeTrapPatch));
             Harmony.CreateAndPatchAll(typeof(RoundManagerPatch));
+            
+            MapObjects.mapObjects.ForEach(m =>
+            {
+                var name = m.mapObject.prefabToSpawn.name;
+                if(name.Contains("TurretContainer") || name.Contains("Landmine") || name.Contains("SpikeRoofTrapHazard")) return;
+                m.mapObject.prefabToSpawn.gameObject.AddComponent<SurprisingHazardSetUp>();
+            });
             
             Logger.LogInfo($"SurprisingHazards is ready!");
         }
@@ -69,13 +81,62 @@ using UnityEngine;
             registeredHazard.networkId = networkBehaviour.NetworkObjectId;
             registeredHazard.surprisingHazardBehavior = surpriseComponent;
         
-            SurprisingHazardsPlugin.instance.RegisteredHazards.TryAdd(networkBehaviour.NetworkObjectId, registeredHazard);
+            instance.RegisteredHazards.TryAdd(networkBehaviour.NetworkObjectId, registeredHazard);
+        }
+        
+        public static void RegisterModdedHazard(NetworkBehaviour networkBehaviour)
+        {
+            var component = Instantiate(instance.surprisingGameObject, Vector3.zero, Quaternion.identity, networkBehaviour.transform);
+            var surpriseComponent = component.GetComponent<SurprisingHazardBehavior>();
+            surpriseComponent.parent = networkBehaviour.gameObject;
+            surpriseComponent.networkId = networkBehaviour.NetworkObjectId;
+            surpriseComponent.isServer = networkBehaviour.IsServer;
+        
+            var registeredHazard = new RegisteredHazard();
+
+            registeredHazard.gameObject = networkBehaviour.gameObject;
+            registeredHazard.networkId = networkBehaviour.NetworkObjectId;
+            registeredHazard.surprisingHazardBehavior = surpriseComponent;
+        
+            instance.RegisteredHazards.TryAdd(networkBehaviour.NetworkObjectId, registeredHazard);
+        }
+
+        public float GetCustomDistance(string name)
+        {
+            float distance = visibleRange.Value;
+            
+            var nameSearch = name.ToLower();
+            while (nameSearch.Contains(" ")) nameSearch = nameSearch.Replace(" ", "");
+            
+            customVisibleRange.Value.Split(",").ToList().ForEach(hazard =>
+            {
+
+                var data = hazard.Split(":");
+                
+                var hazardName = data[0].ToLower();
+                while (hazardName.Contains(" ")) hazardName = hazardName.Replace(" ", "");
+                
+                if (nameSearch.Contains(hazardName))
+                {
+                    distance = float.Parse(data[1]);
+                }
+
+            });
+
+            return distance;
+
         }
 
         private void LoadConfigs()
         {
             visibleRange = Config.Bind("General" ,"visibleRange", 8f, "Hazards visible range. Every player must have the same value, or this will cause desync. No need to restart the game :)");
             CreateFloatConfig(visibleRange);
+            customVisibleRange = Config.Bind("General" ,
+                "customVisibleRange", 
+                "turret:15,spikeRoofTrap:8,landmine:8", 
+                "Hazards visible range. The name must be the hazard prefab name, it works with modded hazards too. Every player must have the same value, or this will cause desync. No need to restart the game :)"
+                );
+            CreateStringConfig(customVisibleRange);
         }
         
         private void CreateFloatConfig(ConfigEntry<float> configEntry, float min = 0f, float max = 100f)