+++ Thunderstore/CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+### v1.0.1
+
+- Add customVisibleRange, where you can custom the visible range for every hazards by using their prefab name
+- Modded hazards will now be affected by the mod
+
 ### v1.0.0
 
 - Release